+++ packages/network/src/config.ts
@@ -121,6 +121,10 @@ export const allNetworks = validateAndSortNetworks([
     disabled: true
   },
   {
+    ...networks[ChainId.POLYGON_AMOY],
+    ...genUrls('amoy')
+  },
+  {
     ...networks[ChainId.BSC_TESTNET],
     ...genUrls('bsc-testnet')
   },
@@ -145,6 +149,10 @@ export const allNetworks = validateAndSortNetworks([
     ...genUrls('homeverse-testnet')
   },
   {
+    ...networks[ChainId.XAI],
+    ...genUrls('xai')
+  },
+  {
     ...networks[ChainId.XAI_SEPOLIA],
     ...genUrls('xai-sepolia')
   },
@@ -157,8 +165,12 @@ export const allNetworks = validateAndSortNetworks([
     ...genUrls('astar-zkevm')
   },
   {
-    ...networks[ChainId.ASTAR_ZKATANA],
-    ...genUrls('astar-zkatana')
+    ...networks[ChainId.ASTAR_ZKYOTO],
+    ...genUrls('astar-zkyoto')
+  },
+  {
+    ...networks[ChainId.XR_SEPOLIA],
+    ...genUrls('xr-sepolia')
   },
   {
     ...networks[ChainId.HARDHAT],