+++ packages/waas/src/index.ts
@@ -4,6 +4,14 @@ export * from './auth'
 export * as store from './store'
 export * as networks from './networks'
 
+export type { Transaction } from './intents/transactions'
+export {
+  erc20,
+  erc721,
+  erc1155,
+  delayedEncode
+} from './intents/transactions'
+
 export * from './intents/responses'
 
 export * as defaults from './defaults'